Library methods reorganization
